+++ src/lib/utils/global.ts
@@ -15,6 +15,7 @@ import { createDataAttributePropertySource } from '../props/property-sources/cre
 import { createJsonScriptPropertySource } from '../props/property-sources/createJsonScriptPropertySource';
 import { createReactivePropertySource } from '../props/property-sources/createReactivePropertySource';
 import { createCustomPropertySource } from '../props/property-sources/createCustomPropertySource';
+import { createFormPropertySource } from '../props/property-sources/createFormPropertySource';
 
 // TODO: Move to "App"?
 class MubanGlobal {
@@ -30,6 +31,7 @@ class MubanGlobal {
     createTextPropertySource(),
     createHtmlPropertySource(),
     createCustomPropertySource(),
+    createFormPropertySource(),
   ];
 }